requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.2.4
streamlit==1.28.2
python-dotenv==1.0.0
pydantic==1.10.13